rename timeline component
