+++ src/components/MainMenu/StagesTimeline.js
